clean tsconfig file & renaming files
